Added certifications API
